refactor : review 도메인으로 이동 (#938)
